+++ src/bunit.web/JSInterop/InvocationHandlers/Implementation/LooseModeJSObjectReferenceInvocationHandler.cs
@@ -1,4 +1,4 @@
-#if !NETSTANDARD2_1
+#if NET5_0_OR_GREATER
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.JSInterop;