+++ src/bunit.web/Rendering/WebTestRenderer.cs
@@ -1,10 +1,10 @@
 using System;
-#if !NETSTANDARD2_1
+#if NET5_0_OR_GREATER
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.DependencyInjection;
 #endif
 using Microsoft.Extensions.Logging;
-#if !NETSTANDARD2_1
+#if NET5_0_OR_GREATER
 using Microsoft.JSInterop;
 #endif
 
@@ -19,12 +19,23 @@ namespace Bunit.Rendering
 		/// <summary>
 		/// Initializes a new instance of the <see cref="WebTestRenderer"/> class.
 		/// </summary>
-		public WebTestRenderer(IRenderedComponentActivator activator, IServiceProvider services, ILoggerFactory loggerFactory)
-			: base(activator, services, loggerFactory)
+		public WebTestRenderer(IRenderedComponentActivator renderedComponentActivator, TestServiceProvider services, ILoggerFactory loggerFactory)
+			: base(renderedComponentActivator, services, loggerFactory)
 		{
-#if !NETSTANDARD2_1
+#if NET5_0_OR_GREATER
 			ElementReferenceContext = new WebElementReferenceContext(services.GetRequiredService<IJSRuntime>());
 #endif
 		}
+
+#if NET5_0_OR_GREATER
+		/// <summary>
+		/// Initializes a new instance of the <see cref="WebTestRenderer"/> class.
+		/// </summary>
+		public WebTestRenderer(IRenderedComponentActivator renderedComponentActivator, TestServiceProvider services, ILoggerFactory loggerFactory, IComponentActivator componentActivator)
+			: base(renderedComponentActivator, services, loggerFactory, componentActivator)
+		{
+			ElementReferenceContext = new WebElementReferenceContext(services.GetRequiredService<IJSRuntime>());
+		}
+#endif
 	}
 }