+++ src/bunit.web/TestContextWrapper.cs
@@ -1,5 +1,6 @@
 using System;
 using Bunit.Rendering;
+using Bunit.TestDoubles;
 using Microsoft.AspNetCore.Components;
 using Microsoft.JSInterop;